+++ my-app/src/redux/slices/charactersSlice.ts
