+++ my-app/src/redux/slices/namesSlice.ts
